% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{package}
\name{aaa.helper.fun-package}
\alias{aaa.helper.fun-package}
\alias{aaa.helper.fun}
\title{TG Therapeutics Corporate Color Palettes}
\description{
Functions for working with TG Therapeutics corporate color palettes
in ggplot2 visualizations.
}
\author{
\strong{Maintainer}: Chris Rowland \email{christopher.r.rowland@gmail.com} (Director, Clinical Data Science)

}
\keyword{internal}
